# api-docs.json
{"openapi":"3.1.0","info":{"title":"Simple To-Do App Backend"},"servers":[{"url":"http://localhost:30002","description":"Generated server url"}],"tags":[{"name":"1 - Usuarios"},{"name":"2 - Canchas"},{"name":"2 - Sessions"},{"name":"3 - Teams"},{"name":"3 - TimeSlots"},{"name":"4 - Bookings"},{"name":"4 - Team Invitations"},{"name":"home-controller"}],"paths":{"/timeslots/field/{fieldId}":{"get":{"tags":["3 - TimeSlots"],"summary":"Listar dias y horarios de servicio de una cancha","operationId":"listTimeSlotsByField","parameters":[{"name":"fieldId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TimeSlotResponseDTO"}}}}}},"security":[{"Bearer Authentication":[]}]},"put":{"tags":["3 - TimeSlots"],"summary":"Reemplazar todos los dias y horarios de servicio de una cancha","operationId":"replaceWeeklySlots","parameters":[{"name":"fieldId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TimeSlotDTO"}}}},"required":true},"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"OK"}},"security":[{"Bearer Authentication":[]}]}},"/timeslots/field/{fieldId}/{dayOfWeek}":{"get":{"tags":["3 - TimeSlots"],"summary":"Listar horarios de servicio de una cancha para un día de semana específico","operationId":"getSlotForDay","parameters":[{"name":"fieldId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"dayOfWeek","in":"path","required":true,"schema":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TimeSlot"}}}}},"security":[{"Bearer Authentication":[]}]},"put":{"tags":["3 - TimeSlots"],"summary":"Reemplazar los horarios de servicio de un día específico sin pisar el resto","operationId":"updateDaySlots","parameters":[{"name":"fieldId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"dayOfWeek","in":"path","required":true,"schema":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSlotDTO"}}},"required":true},"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"OK"}},"security":[{"Bearer Authentication":[]}]}},"/sessions":{"put":{"tags":["2 - Sessions"],"summary":"Refresh a session","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshDTO"}}},"required":true},"responses":{"401":{"description":"Invalid refresh token supplied"},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}},"post":{"tags":["2 - Sessions"],"summary":"Log in, creating a new session","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"401":{"description":"Invalid username or password supplied"},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/fields/{id}":{"put":{"tags":["2 - Canchas"],"summary":"Actualizar cancha","description":"Actualiza los datos de una cancha que pertenece al usuario","operationId":"updateField","parameters":[{"name":"id","in":"path","description":"ID de la cancha","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FieldUpdateDTO"}}},"required":true},"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Cancha actualizada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Field"}}}},"404":{"description":"Cancha no encontrada o no pertenece al usuario","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Field"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Field"}}}}},"security":[{"Bearer Authentication":[]}]},"delete":{"tags":["2 - Canchas"],"summary":"Eliminar cancha (borrado real)","description":"Elimina completamente una cancha de la base de datos","operationId":"deleteField","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"204":{"description":"Cancha eliminada exitosamente"},"404":{"description":"Cancha no encontrada o no pertenece al usuario"}},"security":[{"Bearer Authentication":[]}]}},"/users/register":{"post":{"tags":["1 - Usuarios"],"summary":"Register user","description":"Create a new user account","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDTO"}}},"required":true},"responses":{"201":{"description":"User created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}},"409":{"description":"Some field has a useless information. Maybe information nota provided, an invalid email, an invalid URL.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncorrectValueResponse"}}}},"400":{"description":"Validation errors. The body is a JSON with dynamic keys corresponding to the form fields, for example: { \"username\": \"The username cannot be empty to register a user.\" }","content":{"application/json":{}}}}}},"/users/login":{"post":{"tags":["1 - Usuarios"],"summary":"Login user","description":"Login a user with their credentials to obtain a JWT token","operationId":"login_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"403":{"description":"Unverified email, inactive account, or invalid credentials","content":{"text/plain":{"schema":{"type":"string"}}}},"200":{"description":"Successful logging","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}},"security":[{"Bearer Authentication":[]}]}},"/teams":{"get":{"tags":["3 - Teams"],"summary":"List Teams","description":"Returns all the teams registered in the system","operationId":"getAllTeams","responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Correctly listed equipment","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamDetailsDTO"}}}}},"security":[{"Bearer Authentication":[]}]},"post":{"tags":["3 - Teams"],"summary":"Create team","description":"Allows the authenticated user to create a new team and be its captain.","operationId":"createTeam","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreateDTO"}}},"required":true},"responses":{"403":{"description":"Invalid jwt access token supplied"},"201":{"description":"Team created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamDetailsDTO"}}}},"409":{"description":"Conflict: A team with that name already exists."},"401":{"description":"Invalid Token"}},"security":[{"Bearer Authentication":[]}]}},"/invitations/teams/{teamId}":{"post":{"tags":["4 - Team Invitations"],"summary":"Invite a user to a Team","description":"Allows the captain of the indicated team to send via email an invitation to a user to join its team","operationId":"createInvitation","parameters":[{"name":"teamId","in":"path","description":"ID of the Team","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInvitationDTO"}}},"required":true},"responses":{"403":{"description":"Invalid jwt access token supplied"},"201":{"description":"Invitation created and sent successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvitationDTO"}}}},"404":{"description":"Team indicated was not an existing Team in the system."},"401":{"description":"Unauthorized user: only the captain of the team can invite new members."},"409":{"description":"Conflict: Couldn't send an email to the invitee"}},"security":[{"Bearer Authentication":[]}]}},"/fields":{"post":{"tags":["2 - Canchas"],"summary":"Crear cancha","description":"Permite a un usuario crear una cancha asociada a su cuenta","operationId":"createField","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FieldCreateDTO"}}},"required":true},"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Cancha creada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Field"}}}},"400":{"description":"Datos inválidos o cancha duplicada"}},"security":[{"Bearer Authentication":[]}]}},"/bookings":{"post":{"tags":["4 - Bookings"],"summary":"Create a new booking","description":"Create a new booking for a user, field, date and hour","operationId":"createBooking","parameters":[{"name":"timeslotId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"hour","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"201":{"description":"Booking created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDTO"}}}},"400":{"description":"Invalid user or field"}},"security":[{"Bearer Authentication":[]}]}},"/teams/{id}":{"delete":{"tags":["3 - Teams"],"summary":"Delete Team","description":"Deletes a Team. Only the captain of the team is able to perform this operation.","operationId":"deleteTeam","parameters":[{"name":"id","in":"path","description":"ID of the Team","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Unauthorized. Only the captain is able to delete the team"},"204":{"description":"Team delete successfully"},"404":{"description":"Team Not Found"}},"security":[{"Bearer Authentication":[]}]},"patch":{"tags":["3 - Teams"],"summary":"Update Team","description":"Allows the captain to modify the colors or the logo of the indicated Team.","operationId":"updateTeam","parameters":[{"name":"id","in":"path","description":"ID of the Team","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamUpdateDTO"}}},"required":true},"responses":{"403":{"description":"Unauthorized, only the captain can perform the operation","content":{}},"200":{"description":"Team updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamDetailsDTO"}}}},"404":{"description":"Team not found"}},"security":[{"Bearer Authentication":[]}]}},"/invitations/teams/accept":{"patch":{"tags":["4 - Team Invitations"],"summary":"Accept an invitation to join a team","description":"Allows an invited user to accept an existing and yet pending invitation to join a Team","operationId":"acceptInvitation","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Invitation accepted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamDetailsDTO"}}}},"401":{"description":"Unauthorized: Cant join the team, user legged differs from invited user"},"409":{"description":"Conflict: User was already a member of the team."},"404":{"description":"Not found: The indicated User invited was not found"}},"security":[{"Bearer Authentication":[]}]}},"/fields/{id}/{active}":{"patch":{"tags":["2 - Canchas"],"summary":"Activar /desactivar cancha","description":"Marca la cancha como activa nuevamente","operationId":"setStatusField","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"active","in":"path","required":true,"schema":{"type":"boolean"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Cancha activada correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Field"}}}},"404":{"description":"Cancha no encontrada o no pertenece al usuario","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Field"}}}}},"security":[{"Bearer Authentication":[]}]}},"/users/verify":{"get":{"tags":["1 - Usuarios"],"summary":"Verify user account","description":"Verifies the user account associated with the provided token. The token is sent via email after registration.","operationId":"verifyUser","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Account verified successfully","content":{"text/plain":{"schema":{"type":"string"}}}},"400":{"description":"Token is invalid or expired","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/users/me":{"get":{"tags":["1 - Usuarios"],"summary":"Load user profile","description":"Fetches the profile information of the currently authenticated user.","operationId":"getCurrentUserProfile","responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"User profile successfully loaded","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileDTO"}}}},"401":{"description":"User is not authenticated or session has expired","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"Bearer Authentication":[]}]}},"/teams/my-teams":{"get":{"tags":["3 - Teams"],"summary":"List teams where the user is member","description":"Returns a list of the teams (TeamDetailsDTO) where the user is a member","operationId":"getUsersTeams","responses":{"403":{"description":"UserNotFound, invalid session"},"200":{"description":"Teams listed successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TeamDetailsDTO"}}}}}},"security":[{"Bearer Authentication":[]}]}},"/invitations/teams/{teamId}/pending":{"get":{"tags":["4 - Team Invitations"],"summary":"get pending invitations to a Team","description":"Allows the captain of the indicated team to get the pending invitations that have been to join the indicated team","operationId":"getPendingInvitation","parameters":[{"name":"teamId","in":"path","description":"ID of the Team","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Invitations listed successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InvitationDTO"}}}}},"404":{"description":"Team indicated was not an existing Team in the system."},"401":{"description":"Unauthorized user: Only the captain can to the pending invitations to a team"}},"security":[{"Bearer Authentication":[]}]}},"/fields/owner/summary":{"get":{"tags":["2 - Canchas"],"summary":"Resumen para dueño de canchas","description":"Cantidad de canchas, reservas de hoy y porcentaje de ocupación","operationId":"getOwnerSummary","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerSummaryDTO"}}}}},"security":[{"Bearer Authentication":[]}]}},"/fields/own":{"get":{"tags":["2 - Canchas"],"summary":"Listar canchas propias","description":"Obtiene todas las canchas asociadas al usuario actual","operationId":"listMyFields","responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Listado de canchas obtenido exitosamente","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Field"}}}}}},"security":[{"Bearer Authentication":[]}]}},"/fields/all":{"get":{"tags":["2 - Canchas"],"summary":"Listar todas las canchas activas","description":"Devuelve las canchas disponibles para cualquier usuario","operationId":"listAllActiveFields","responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Listado de canchas activas","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Field"}}}}}},"security":[{"Bearer Authentication":[]}]}},"/bookings/{id}":{"get":{"tags":["4 - Bookings"],"summary":"Get booking by ID","description":"Returns the booking details by booking ID","operationId":"getBookingById","parameters":[{"name":"id","in":"path","description":"ID of the booking","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"Booking found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDTO"}}}},"404":{"description":"Booking not found"}},"security":[{"Bearer Authentication":[]}]},"delete":{"tags":["4 - Bookings"],"summary":"Cancel a booking","description":"Marks a booking as inactive (cancelled) without deleting it","operationId":"cancelBooking","parameters":[{"name":"id","in":"path","description":"ID of the booking to cancel","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"204":{"description":"Booking cancelled successfully"},"404":{"description":"Booking not found"}},"security":[{"Bearer Authentication":[]}]}},"/bookings/owner":{"get":{"tags":["4 - Bookings"],"summary":"Reservas sobre canchas del dueño logueado","operationId":"getBookingsForOwner","responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookingDTO"}}}}}},"security":[{"Bearer Authentication":[]}]}},"/bookings/my":{"get":{"tags":["4 - Bookings"],"summary":"Reservas propias del usuario logueado","operationId":"getMyBookings","responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookingDTO"}}}}}},"security":[{"Bearer Authentication":[]}]}},"/bookings/field/{fieldId}":{"get":{"tags":["4 - Bookings"],"summary":"Get active bookings by field","description":"Returns all active bookings for a specific field","operationId":"getBookingsByField","parameters":[{"name":"fieldId","in":"path","description":"ID of the field","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"List of active bookings for the field","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDTO"}}}},"404":{"description":"Field not found"}},"security":[{"Bearer Authentication":[]}]}},"/bookings/availability/{fieldId}":{"get":{"tags":["4 - Bookings"],"summary":"Horas disponibles para reservar","description":"Devuelve las horas libres por día para los próximos X días","operationId":"getAvailableHours","parameters":[{"name":"fieldId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"403":{"description":"Invalid jwt access token supplied"},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"integer","format":"int32"}}}}}}},"security":[{"Bearer Authentication":[]}]}},"/":{"get":{"tags":["home-controller"],"operationId":"home","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"TimeSlotDTO":{"type":"object","properties":{"dayOfWeek":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"openTime":{"type":"integer","format":"int32","maximum":23,"minimum":0},"closeTime":{"type":"integer","format":"int32","maximum":24,"minimum":1}}},"RefreshDTO":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"TokenDTO":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken"]},"FieldUpdateDTO":{"type":"object","properties":{"name":{"type":"string"},"grassType":{"type":"string"},"lighting":{"type":"boolean"},"zone":{"type":"string"},"address":{"type":"string"},"photoUrl":{"type":"string"},"price":{"type":"number","format":"double"}},"required":["address","grassType","lighting","name","photoUrl","price","zone"]},"Field":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"grassType":{"type":"string"},"lighting":{"type":"boolean"},"zone":{"type":"string"},"address":{"type":"string"},"photoUrl":{"type":"string"},"price":{"type":"number","format":"double"},"active":{"type":"boolean"},"owner":{"$ref":"#/components/schemas/User"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"password":{"type":"string"},"role":{"type":"string"},"username":{"type":"string"},"emailVerified":{"type":"boolean"},"active":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"enabled":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"}}},"UserCreateDTO":{"type":"object","description":"User registration data","properties":{"name":{"type":"string","description":"User's name"},"last_name":{"type":"string","description":"User's last name"},"username":{"type":"string","description":"Username address"},"password":{"type":"string","description":"User's password"},"role":{"type":"string","description":"User's role"},"gender":{"type":"string","description":"User's gender"},"age":{"type":"string","description":"User's age"},"zone":{"type":"string","description":"User's zone"},"urlProfilePicture":{"type":"string","description":"User profile picture"}},"required":["age","gender","last_name","name","password","role","urlProfilePicture","username","zone"]},"IncorrectValueResponse":{"type":"object","properties":{"field":{"type":"string"},"error_description":{"type":"string"}}},"UserLoginDTO":{"type":"object","properties":{"username":{"type":"string","description":"User's name"},"password":{"type":"string","description":"User's password"}},"required":["password","username"]},"TeamCreateDTO":{"type":"object","properties":{"name":{"type":"string"},"primaryColor":{"type":"string"},"secondaryColor":{"type":"string"},"logo":{"type":"string"}},"required":["name"]},"TeamDetailsDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"captain":{"type":"string"},"primaryColor":{"type":"string"},"secondaryColor":{"type":"string"},"logo":{"type":"string"},"membersUsernames":{"type":"array","items":{"type":"string"}}}},"CreateInvitationDTO":{"type":"object","properties":{"inviteeEmail":{"type":"string"}}},"InvitationDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"team_id":{"type":"integer","format":"int64"},"inviteeEmail":{"type":"string"},"pending":{"type":"boolean"},"token":{"type":"string"}}},"FieldCreateDTO":{"type":"object","description":"Datos de la cancha","properties":{"name":{"type":"string"},"grassType":{"type":"string"},"lighting":{"type":"boolean"},"zone":{"type":"string"},"address":{"type":"string"},"photoUrl":{"type":"string"},"price":{"type":"number","format":"double"}},"required":["address","grassType","lighting","name","photoUrl","price","zone"]},"BookingDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"timeSlotId":{"type":"integer","format":"int64"},"bookingDate":{"type":"string","format":"date"},"bookingHour":{"type":"integer","format":"int32"},"active":{"type":"boolean"}},"required":["bookingDate","bookingHour","id","timeSlotId","userId"]},"TeamUpdateDTO":{"type":"object","properties":{"name":{"type":"string"},"primaryColor":{"type":"string"},"secondaryColor":{"type":"string"},"logo":{"type":"string"}}},"UserProfileDTO":{"type":"object","description":"User profile information","properties":{"email":{"type":"string","description":"User email address"},"name":{"type":"string","description":"User first name"},"last_name":{"type":"string","description":"User last name"},"age":{"type":"string","description":"User age"},"gender":{"type":"string","description":"User gender"},"zone":{"type":"string","description":"User zone"},"role":{"type":"string","description":"User role"},"profile_picture":{"type":"string","description":"User profile picture"}},"required":["age","email","gender","last_name","name","profile_picture","role","zone"]},"ErrorResponse":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"}}},"TimeSlotResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dayOfWeek":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"openTime":{"type":"integer","format":"int32"},"closeTime":{"type":"integer","format":"int32"},"fieldId":{"type":"integer","format":"int64"}}},"TimeSlot":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dayOfWeek":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]},"openTime":{"type":"integer","format":"int32"},"closeTime":{"type":"integer","format":"int32"},"field":{"$ref":"#/components/schemas/Field"}}},"OwnerSummaryDTO":{"type":"object","properties":{"totalFields":{"type":"integer","format":"int32"},"totalBookingsToday":{"type":"integer","format":"int32"},"occupancyPercentage":{"type":"number","format":"double"}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}